Web hotfix (#65)
## Problem
Web container failed to build during the workflow run. 

## ESLint Error
```
#12 1.279 info  - Linting and checking validity of types...
#12 1.689 error - ESLint: ESLint configuration in .eslintrc.json is invalid: 	- Unexpected top-level property "[javascript]". 
#12 10.10 Failed to compile.
```

### Solution
The `.eslintrc.json` configuration file was incorrectly formatted. It hadn't changed during the typescript conversion, but the package for `eslint` did get updated. Since this issue did not appear in previous build runs either how eslint interprets the file changed during updates or it wasn't previously used. Either way, I was able to reproduce the error in my local dev environment by running `yarn build` in the `app/web` directory. I got a working config file created, and now that it is working I had to fix the issues in the `.ts/.tsx` files the linter found (it turns our, ESLint was not installed in my Ubuntu IDE, so the default typescript linter was more lax).

I am also going to suggest creating a devcontainer in the near future for web development so that we sync the dev and prod eslints.


## Missing Property Error
```
#12 10.10 ./utils/constants.ts:3:14
#12 10.10 Type error: Property 'patient_encounter_uuid' is missing in type '{ patient_rfid: string; document_num: string; location: string; handover_from: string; arrival_date: Date; arrival_time: Date; triage_acuity: string; age: number; gender: string; ... 8 more ...; comment: string; }' but required in type 'PatientEncounterFormDataInterface'.
#12 10.10 
#12 10.10   1 | import { PatientEncounterFormDataInterface } from "../interfaces/PatientEncounterFormDataInterface";
#12 10.10   2 | 
#12 10.10 > 3 | export const MFPEFormData: PatientEncounterFormDataInterface = {
#12 10.10     |              ^
#12 10.10   4 |   patient_rfid: "",
#12 10.10   5 |   document_num: "",
#12 10.10   6 |   location: "Main Medical",
```

### Solution
* Added missing type to PatientEncounterFormDataInterface 

## Ticket URL
https://mediform.atlassian.net/browse/MEDI-17

## Documentation
N/A

## Tests Run
* Tested locally. Rebuilt `web` image and container from scratch
* Manually tested web application
    - Login
    - Create a new patient encounter form
    - Search PE forms
    - View existing PE
    - Update existing PEs
    - Delete PEs